Added full-sprint-qualifying.ts
