+++ src/scraper/full-sprint-qualifying.ts
@@ -1,0 +1,22 @@
+import { staticLinks } from "../endpoints/endpoints";
+
+import { isFullRaceResult } from "../types/types";
+import { assignQualiValues, getF1Table, getResultURL } from "../utils/scrapping";
+
+/**
+ *
+ * @param {year} year - 1950-now
+ * @param {string} [raceName="Australia"]
+ * @returns {Promise<isFullRaceResult[]>}
+ */
+
+export const getFullSprintShootout = async (year: number = new Date().getFullYear(), raceName: string = "Brazil"): Promise<isFullRaceResult[]> => {
+    try {
+        const resultsURL = await getResultURL(year, raceName);
+        const sprintQualiURL = `${staticLinks.fullResults}/${year}/${resultsURL.slice(23, resultsURL.length - 11)}/sprint-qualifying`;
+
+        return getF1Table(sprintQualiURL, assignQualiValues) as unknown as isFullRaceResult[];
+    } catch (error: unknown) {
+        throw new Error(error as string);
+    }
+};
