+++ src/server.ts
@@ -12,3 +12,4 @@ export { getFullFastestLaps } from "./scraper/full-fastest-laps";
 export { getFullPracticeResults } from "./scraper/full-practice-results";
 export { getFullPitstopsSummary } from "./scraper/full-pitstops-summary";
 export { getStartingGrid } from "./scraper/starting-grid";
+export { getFullSprintShootout } from "./scraper/full-sprint-qualifying";
