+++ src/utils/scrapping.ts
@@ -1,8 +1,16 @@
 import axios from "axios";
 import { dynamicLinks } from "../endpoints/endpoints";
 import * as cheerio from "cheerio";
+import { fullQualiResult, qualiTimes } from "../types/types";
+import { assignPropertyIfDefined } from "./common";
 
-export async function getResultURL(year: number = new Date().getFullYear(), raceName: string) {
+/**
+ *
+ * @param {number} year 1950-now
+ * @param {string} raceName
+ * @returns {Promise<string>}
+ */
+export async function getResultURL(year: number = new Date().getFullYear(), raceName: string): Promise<string> {
     try {
         const response = await axios(`${dynamicLinks.rootLink}/${year}/${dynamicLinks.results}`);
 
@@ -19,6 +27,12 @@ export async function getResultURL(year: number = new Date().getFullYear(), race
     }
 }
 
+/**
+ *
+ * @param {string} url
+ * @param {(driver: string[]) => object}callback
+ * @returns {Promise<Array<object>}
+ */
 export async function getF1Table(url: string, callback: (driver: string[]) => object): Promise<Array<object>> {
     try {
         const results = await axios(url);
@@ -38,4 +52,23 @@ export async function getF1Table(url: string, callback: (driver: string[]) => ob
     } catch (error: unknown) {
         throw new Error(error as string);
     }
+}
+
+/**
+ *
+ * @param {string} driver - values from table
+ * @returns {fullQualiResult}
+ */
+export function assignQualiValues(driver: string[]): fullQualiResult {
+    const driverDetails: fullQualiResult = {
+        position: driver[0],
+        number: parseInt(driver[1]),
+        name: driver[2].slice(0, driver[2].length - 3).replace(/\u00a0/g, " "),
+        code: driver[2].slice(driver[2].length - 3),
+        team: driver[3],
+        times: assignPropertyIfDefined([driver[4], driver.length > 6 ? driver[5] : false, driver.length > 7 ? driver[6] : false], ["q1", "q2", "q3"]) as qualiTimes,
+        laps: parseInt(driver[driver.length - 1]),
+    };
+
+    return driverDetails;
 }
